Add fresh commit for Pepijn's version-tar branch.
This code was taken from the latest version of this repo / pull request:

https://github.com/pepijndevos/onepercentsite/tree/versioned-tar

I updated the code to be compatible with the latest version of the
onepercentclub-site repo but I didn't update the code to make it work on
production.

A fresh commit was required because the history in Pepijn's branch
against onepercentsite-old (the former onepercentsite repo) is not
compatible with the onepercentclub-site repo (the former bluebottle
repo).